// client/src/app/company/types/jobForm.ts
export interface JobFormData {
    title: string
    description: string
    salary: string
    location: string
    role: string
    responsibilities: string
    qualifications: string
    job_type: string
    work_location: string
    openings: string
    is_active: boolean
    skills_required: string[]
}